+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - introducing `withLock` wrapper
 - introducing `join` command
+- introducing `sum` command
 - improved error messages by using command name prefix (fix #88)
 - multiline pipelines (fix #83)